Adding more places where cache expiration syncs with filesystem/other caches
Also, fix broken CDI provider
