+++ addons/koji/common/src/main/java/org/commonjava/indy/koji/content/KojiMavenMetadataProvider.java
@@ -36,15 +36,17 @@ import org.commonjava.cdi.util.weft.Locker;
 import org.commonjava.cdi.util.weft.WeftExecutorService;
 import org.commonjava.cdi.util.weft.WeftManaged;
 import org.commonjava.indy.IndyWorkflowException;
-import org.commonjava.indy.core.change.ContentCleanupHelper;
+import org.commonjava.indy.content.DirectContentAccess;
+import org.commonjava.indy.core.content.group.GroupMergeHelper;
 import org.commonjava.indy.data.IndyDataException;
 import org.commonjava.indy.data.StoreDataManager;
 import org.commonjava.indy.koji.conf.IndyKojiConfig;
 import org.commonjava.indy.koji.inject.KojiMavenVersionMetadataCache;
 import org.commonjava.indy.koji.inject.KojiMavenVersionMetadataLocks;
 import org.commonjava.indy.measure.annotation.Measure;
-import org.commonjava.indy.model.core.ArtifactStore;
+import org.commonjava.indy.model.core.Group;
 import org.commonjava.indy.model.core.StoreKey;
+import org.commonjava.indy.pkg.maven.content.MetadataCacheManager;
 import org.commonjava.indy.pkg.maven.content.MetadataInfo;
 import org.commonjava.indy.pkg.maven.content.MetadataKey;
 import org.commonjava.indy.pkg.maven.content.cache.MavenMetadataCache;
@@ -52,9 +54,13 @@ import org.commonjava.indy.pkg.maven.content.cache.MavenMetadataKeyCache;
 import org.commonjava.indy.pkg.maven.content.group.MavenMetadataProvider;
 import org.commonjava.indy.subsys.infinispan.CacheHandle;
 import org.commonjava.indy.subsys.infinispan.CacheProducer;
+import org.commonjava.indy.util.LocationUtils;
 import org.commonjava.maven.galley.TransferException;
 import org.commonjava.maven.galley.event.EventMetadata;
 import org.commonjava.maven.galley.maven.util.ArtifactPathUtils;
+import org.commonjava.maven.galley.model.ConcreteResource;
+import org.commonjava.maven.galley.model.Transfer;
+import org.commonjava.maven.galley.spi.nfc.NotFoundCache;
 import org.infinispan.commons.util.concurrent.ConcurrentHashSet;
 import org.infinispan.manager.DefaultCacheManager;
 import org.infinispan.notifications.Listener;
@@ -67,6 +73,7 @@ import javax.annotation.PostConstruct;
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 import java.io.File;
+import java.io.IOException;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Calendar;
@@ -100,23 +107,23 @@ public class KojiMavenMetadataProvider
     private static final java.lang.String LAST_UPDATED_FORMAT = "yyyyMMddHHmmss";
 
     @Inject
-    @MavenMetadataCache
-    private CacheHandle<MetadataKey, MetadataInfo> metadataCache;
-
-    @Inject
-    @MavenMetadataKeyCache
-    private CacheHandle<MetadataKey, MetadataKey> metadataKeyCache;
+    private MetadataCacheManager mavenMetadataCaches;
 
     @Inject
     private StoreDataManager storeDataManager;
 
     @Inject
-    private ContentCleanupHelper contentCleanupHelper;
+    private GroupMergeHelper helper;
+
+    @Inject
+    private DirectContentAccess fileManager;
+
+    @Inject
+    private NotFoundCache nfc;
 
     @Inject
     @KojiMavenVersionMetadataCache
     private CacheHandle<ProjectRef, Metadata> versionMetadata;
-
 
     @Inject
     private IndyKojiContentProvider kojiContentProvider;
@@ -171,18 +178,19 @@ public class KojiMavenMetadataProvider
         logger.info( "Koji metadata expired for GA: {}", e.getKey() );
         try
         {
-            Set<StoreKey> keys = storeDataManager.query()
-                                                    .getAll( s -> group == s.getType() && kojiConfig.isEnabledFor(
-                                                            s.getName() ) )
-                                                    .stream()
-                                                    .map( s -> s.getKey() )
-                                                    .collect( Collectors.toSet() );
+            List<Group> affected = storeDataManager.query()
+                                                           .getAll(
+                                                                   s -> group == s.getType() && kojiConfig.isEnabledFor(
+                                                                           s.getName() ) )
+                                                           .stream()
+                                                           .map( s -> (Group) s )
+                                                           .collect( Collectors.toList() );
 
-            if ( !keys.isEmpty() )
+            if ( !affected.isEmpty() )
             {
-                logger.info( "Triggering metadata cleanup from Koji metadata expiration, for GA: {} in groups: {}", e.getKey(), keys );
+                logger.info( "Triggering metadata cleanup from Koji metadata expiration, for GA: {} in groups: {}", e.getKey(), affected );
                 String path = ArtifactPathUtils.formatMetadataPath( e.getKey(), METADATA_NAME );
-                contentCleanupHelper.clearPaths( keys, s -> path.equals( s ), true );
+                clearPaths( affected, path );
             }
 
         }
@@ -194,6 +202,44 @@ public class KojiMavenMetadataProvider
         {
             logger.error( "Failed to format metadata path for: " + e.getKey(), ex );
         }
+    }
+
+    private void clearPaths( final List<Group> affected, final String path )
+    {
+        Logger logger = LoggerFactory.getLogger( getClass() );
+
+        affected.forEach( group->{
+            try
+            {
+                // delete so it'll be recomputed.
+                final Transfer target = fileManager.getTransfer( group, path );
+
+                if ( target.exists() )
+                {
+                    logger.debug( "Deleting merged file: {}", target );
+                    target.delete( false );
+                    if ( target.exists() )
+                    {
+                        logger.error( "\n\n\n\nDID NOT DELETE merged metadata file at: {} in group: {}\n\n\n\n", path,
+                                      group.getName() );
+                    }
+                    helper.deleteChecksumsAndMergeInfo( group, path );
+                }
+                else
+                {
+                    ConcreteResource resource = new ConcreteResource( LocationUtils.toLocation( group ), path );
+                    nfc.clearMissing( resource );
+                }
+
+                // make sure we delete these, even if they're left over.
+                helper.deleteChecksumsAndMergeInfo( group, path );
+            }
+            catch ( final IndyWorkflowException | IOException e )
+            {
+                logger.error( "Failed to delete generated file (to allow re-generation on demand: {}/{}. Error: {}", e,
+                              group.getKey(), path, e.getMessage() );
+            }
+        } );
     }
 
     @Override
